Code Refactoring Repository Lifecycle Handled
